% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{blankBC}
\alias{blankBC}
\title{Blank Budyko Curve Graph}
\format{An object of class \code{gg} (inherits from \code{ggplot}) of length 9.}
\usage{
blankBC
}
\value{
Generates a blank Budyko Curve with energy and water limit
lines
}
\description{
Blank Budyko Curve Graph
}
\examples{
library(ggplot2)
blankBC=ggplot(data=data.frame(PET.P=seq(0,20,1),AET.P=c(0,rep(1,20))),aes(x=PET.P,y=AET.P))+
geom_line()+xlab("Aridity Index")+ylab("Evaporative Index")
blankBC+coord_cartesian(xlim=c(0,2))+geom_vline(xintercept=1,lty=2)+
    geom_text(data=data.frame(
        labels=c("Energy Limited","Water Limited"),
        x=c(0.7,1.3),y=c(0.25,0.25)
        ),aes(x=x,y=y,label=labels))
fit=as.data.frame(c("param"=2,"err"=0.5,"hshift"=0))
blankBC+
    geom_line(data=budyko_sim(fit=fit,method="Fu"))
}
\keyword{datasets}
