% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{budyko_sim}
\alias{budyko_sim}
\title{Simulate Budyko Curve}
\usage{
budyko_sim(
  p = 2,
  fit = NULL,
  method = "budyko",
  res = 0.01,
  hshift = FALSE,
  hs = NULL,
  silent = FALSE
)
}
\arguments{
\item{p}{\link{numeric} parameter value for custom Budyko curve}

\item{fit}{\link{list} dataframe object with names() method and entries
"param" for parameter p, "err" for vertical fit error, and "hshift" for
optional hshift value (default to zero)}

\item{method}{A string. One of the following: Fu, Turc-Pike, Wang-Tank}

\item{res}{\link{numeric} resolution of fit, defaults to 0.01 units}

\item{hshift}{A boolean.  T/F (default F), whether to test for horizontal shift}

\item{hs}{\link{numeric} optional horizontal shift value}

\item{silent}{A boolean.  (default F) display error message?}
}
\value{
Given a Budyko Curve type and fit parameter,
simulates a smooth curve in Budyko-space
}
\description{
Simulate Budyko Curve
}
\examples{
library(ggplot2)
fit=as.data.frame(c("param"=2,"err"=0.5,"hshift"=0))
blankBC+
    geom_line(data=budyko_sim(fit=fit,method="Fu"))+
    coord_cartesian(xlim=c(0,5))+
    theme_bw()
}
