% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Simulated Data with Fu Budyko omega=2}
\format{An object of class \code{data.frame} with 30 rows and 3 columns.}
\usage{
testdata
}
\value{
20 years of simulated data from a Budyko Fu formulation
and omega parameter value equal to 2 and average vertical fit error equal
to 0.05
}
\description{
Simulated Data with Fu Budyko omega=2
}
\keyword{datasets}
