% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{budyko_errbounds}
\alias{budyko_errbounds}
\title{Generate errorbounds from a Budyko fit}
\usage{
budyko_errbounds(fit, res = 0.1, hshift = FALSE)
}
\arguments{
\item{fit}{\link{list} dataframe object with names() method and entries
"param" for parameter p, "err" for vertical fit error, and "hshift" for
optional hshift value (default to zero)}

\item{res}{\link{numeric} resolution of fit, defaults to 0.01 units}

\item{hshift}{A boolean.  T/F (default F), whether to test for horizontal shift.
hs value must be defined in fit}
}
\value{
Given a "fit" dataframe object from budyko::budyko_fit,
generate a dataframe of error bounds that can be fed to plot or
ggplot2::ggplot. name (as in names()) of fit must be the method
}
\description{
Generate errorbounds from a Budyko fit
}
\examples{
library(ggplot2)
fit=budyko_fit(data=testdata,method="Fu")
blankBC+
  geom_polygon(data=budyko_errbounds(fit),fill="gray")+
  geom_line(data=budyko_sim(fit=fit))+
  geom_point(data=testdata)+
  coord_cartesian(xlim=c(0,5))
}
